enhance: introduce new list view for microlearnings
