+++ apps/frontend-manage/src/pages/courses/[id].tsx
@@ -1,4 +1,5 @@
 import { useQuery } from '@apollo/client'
+import MicroLearningList from '@components/courses/MicroLearningList'
 import PracticeQuizList from '@components/courses/PracticeQuizList'
 import { faCrown } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
@@ -20,7 +21,6 @@ import CourseOverviewHeader from '../../components/courses/CourseOverviewHeader'
 import CourseSettings from '../../components/courses/CourseSettings'
 import GroupActivityTile from '../../components/courses/GroupActivityTile'
 import LiveSessionList from '../../components/courses/LiveSessionList'
-import MicroSessionTile from '../../components/courses/MicroSessionTile'
 
 function CourseOverviewPage() {
   const t = useTranslations()
@@ -158,40 +158,23 @@ function CourseOverviewPage() {
               <Tabs.TabContent
                 key="content-microlearnings"
                 value="microLearnings"
+                className={{ root: 'px-2 py-2' }}
               >
-                MICROLEARNINGS
+                <MicroLearningList
+                  microLearnings={course.microLearnings ?? []}
+                  userCatalyst={user?.userProfile?.catalyst}
+                />
               </Tabs.TabContent>
               <Tabs.TabContent
                 key="content-groupActivities"
                 value="groupActivities"
+                className={{ root: 'px-2 py-2' }}
               >
                 GROUP ACTIVITIES
               </Tabs.TabContent>
             </Tabs>
           </div>
           <div>
-            <div className="mb-4">
-              <H3 className={{ root: 'flex flex-row gap-3' }}>
-                <div>{t('shared.generic.microlearnings')}</div>
-                <Button.Icon className={{ root: 'text-orange-400' }}>
-                  <FontAwesomeIcon icon={faCrown} size="sm" />
-                </Button.Icon>
-              </H3>
-              {course.microSessions && course.microSessions.length > 0 ? (
-                <div className="flex flex-col gap-2 pr-4 overflow-x-auto sm:flex-row">
-                  {course.microSessions.map((microSession) => (
-                    <MicroSessionTile
-                      microSession={microSession}
-                      key={microSession.id}
-                    />
-                  ))}
-                </div>
-              ) : user?.userProfile?.catalyst ? (
-                <div>{t('manage.course.noMicrolearnings')}</div>
-              ) : (
-                <CatalystNotification />
-              )}
-            </div>
             <div className="mb-4">
               <H3 className={{ root: 'flex flex-row gap-3' }}>
                 <div>{t('shared.generic.groupActivities')}</div>
